refactor(openhuman): codify canonical module shape and tidy offenders
Document the canonical per-domain module shape (mod/types/store/ops/
schemas/tools/bus/tests) in CLAUDE.md as the single source of truth, with
an AGENTS.md pointer. Then fix the genuine structural offenders to match:

- memory_queue: move the backfill-progress static + fns and
  ensure_reembed_backfill out of mod.rs into a new ops.rs; mod.rs
  re-exports them so all caller paths stay stable.
- cwd_jail: move the stray `impl Jail { canonicalize_or_log }` into the
  existing impl in jail.rs; keep the spawn/default_backend facade in
  mod.rs under the documented thin-facade exception.
- embeddings, integrations: relocate large inline test suites to sibling
  mod_tests.rs via #[path], leaving both mod.rs bodies as pure re-exports.

No public API changes; verified with cargo check, cargo fmt --check, and
targeted cargo tests for each touched module.
